moved text widgets to their own sub dir
