+++ lib/widgets/misc/bottom_back_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'body_text.dart';
+import 'package:discover_deep_cove/widgets/misc/text/body_text.dart';
 
 class BottomBackButton extends StatelessWidget {
   @override
