+++ lib/widgets/misc/loading_modal_overlay.dart
@@ -1,4 +1,4 @@
-import 'package:discover_deep_cove/widgets/misc/body_text.dart';
+import 'package:discover_deep_cove/widgets/misc/text/body_text.dart';
 import 'package:flutter/material.dart';
 
 class LoadingModalOverlay extends StatelessWidget {
